+++ cmd/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"os"
+	"weather"
+)
+
+func main() {
+	weather.RunCLI(os.Args[1:], os.Stdout)
+}
